Validaciones

+++ src/main/java/OrderlyAPI/Expo2025/Models/DTO/PlatilloDTO.java
@@ -9,6 +9,6 @@ public class PlatilloDTO {
     private Long Id;
     private String NomPlatillo;
     private String Descripcion;
-    private Long Precio;
+    private double Precio;
     private Long TiempoPreparacion;
 }
